requirements.txt
pyodbc
python-dotenv
Pillow
html2image
num2words